requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.2
django-cors-headers==4.5.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
gTTS==2.5.3
idna==3.10
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
